feat: create single commit across all repositories
When a particular flag is sent to the commit operation (right now it's
"-F same", which needs to be changed to something better) open an
editor window using $EDITOR, and use that single commit message for
all subrepositories.

The mechanism is to use `git commit`'s -F option, but replace the
"same" value with the path to the temporary file opened in the editor.